+++ src/reanimated2/UpdateProps.js
@@ -1,8 +1,29 @@
-import { _updatePropsJS } from './js-reanimated/index';
 import {
   updatePropsProcessColors,
   getUpdateProps,
 } from './platform-specific/PlatformSpecific';
+import { processColor } from './Colors';
+import { makeShareable } from './core';
+
+// copied from react-native/Libraries/Components/View/ReactNativeStyleAttributes
+const colorProps = [
+  'backgroundColor',
+  'borderBottomColor',
+  'borderColor',
+  'borderLeftColor',
+  'borderRightColor',
+  'borderTopColor',
+  'borderStartColor',
+  'borderEndColor',
+  'color',
+  'shadowColor',
+  'textDecorationColor',
+  'tintColor',
+  'textShadowColor',
+  'overlayColor',
+];
+
+const ColorProperties = makeShareable(colorProps);
 
 export default function updateProps(
   viewDescriptor,
@@ -20,8 +41,8 @@ export default function updateProps(
     });
   }
 
-  updatePropsProcessColors(updates);
-
+  updatePropsProcessColors(updates, ColorProperties, processColor);
+  
   getUpdateProps()(
     viewDescriptor.value.tag,
     viewName,
